refactor: filter 에 대한 타입 분리 및 다른필드도 필터링 가능하도록 수정
