+++ src/entities/filter/models/index.d.ts
@@ -1,0 +1,2 @@
+export type SortKey = "createdAt";
+export type SortOrder = "asc" | "desc";
