+++ src/entities/filter/ui/PageInfo.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { usePageInfo } from "@/entities/pagination/hooks/usePageInfo";
+import { usePageInfo } from "@/entities/filter/hooks/usePageInfo";
 import { cn } from "@/shared/lib/utils";
 import { Button } from "@/shared/ui/button";
 import {
